+++ src/LocationsServiceProvider.php
@@ -1,14 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tipoff\Locations;
 
-use Illuminate\Support\Str;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
-use Tipoff\Locations\Commands\LocationsCommand;
 
 class LocationsServiceProvider extends PackageServiceProvider
 {
+    public function boot()
+    {
+        $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
+        parent::boot();
+    }
+
     public function configurePackage(Package $package): void
     {
         /*
@@ -18,27 +24,6 @@ class LocationsServiceProvider extends PackageServiceProvider
          */
         $package
             ->name('locations')
-            ->hasConfigFile()
-            ->hasViews()
-            ->hasMigrations([
-                '2020_02_17_170000_create_markets_table',
-                '2020_02_17_180000_create_locations_table',
-            ])
-            ->hasCommand(LocationsCommand::class);
-    }
-
-    /**
-     * Using packageBooted lifecycle hooks to override the migration file name.
-     * We want to keep the old filename for now.
-     */
-    public function packageBooted()
-    {
-        foreach ($this->package->migrationFileNames as $migrationFileName) {
-            if (! $this->migrationFileExists($migrationFileName)) {
-                $this->publishes([
-                    $this->package->basePath("/../database/migrations/{$migrationFileName}.php.stub") => database_path('migrations/' . Str::finish($migrationFileName, '.php')),
-                ], "{$this->package->name}-migrations");
-            }
-        }
+            ->hasConfigFile();
     }
 }