+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `locations` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2021-02-09
 
 - initial release
+
+## 1.1.0 - 2021-02-10
+
+- migrations loaded via package
+- model resolution via container